requirements.txt
# Fleet Gateway Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# GraphQL
strawberry-graphql[fastapi]>=0.216.0

# Database & Caching
redis>=5.0.0
supabase>=2.0.0

# ROS Bridge
roslibpy>=1.5.0

# Configuration (Phase 1)
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0

# Development
python-dotenv>=1.0.0
